makes oneshot corrections adjust ComicInfo
For whatever reason, adjusting the series title, title sort, and book title in the database doesn't make a oneshot searchable based on those fields. Somewhere in the database there's still a remnant of their old name, and that's what Komga searches for when importing a reading list. The only solution I think that's left for me is to change the ComicInfo.xml itself, and have Komga use that import the books with the right title in the first place.